Extract fixtures from dnf_config_manager tests
