+++ tests/unit/plugins/modules/dnf_config_manager/fixtures/mock_dnf4_repolist_status_before_id.txt
@@ -1,0 +1,3 @@
+Repo-id            : appstream-debuginfo
+Repo-status        : disabled
+Repo-status        : disabled
